shareable ticket WIP
