+++ src/routes/init/ticket/+page.ts
@@ -1,8 +1,8 @@
 import { redirect } from '@sveltejs/kit';
+import { isLoggedInGithub } from '../helpers.js';
 
-export const load = async ({ parent }) => {
-    const data = await parent();
-    if (data.ticket) {
-        throw redirect(307, '/init/ticket/thank-you');
+export const load = async () => {
+    if (await isLoggedInGithub()) {
+        throw redirect(307, '/init/ticket/customize');
     }
 };
