--- src/routes/init/ticket/thank-you/+page.ts
@@ -1,6 +1,0 @@
-import { redirect } from '@sveltejs/kit';
-
-export const load = async ({ parent }) => {
-    const data = await parent();
-    if (!data.ticket) throw redirect(307, '/init/ticket');
-};
